requirements.txt
fastapi==0.116.0
uvicorn==0.35.0
sqlalchemy==2.0.41
alembic==1.16.4
psycopg2-binary==2.9.10
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.20
redis==6.2.0
celery==5.5.3
pandas==2.3.1
scikit-learn==1.7.0
python-dotenv==1.0.1
httpx==0.27.2
pytest==8.3.4
pytest-asyncio==0.24.0
# Email dependencies
aiosmtplib==3.0.1
email-validator==2.1.0.post1
jinja2==3.1.3
# 2FA dependencies
pyotp==2.8.0
qrcode==7.4.2
pillow==10.2.0
# Security improvements
python-magic==0.4.27
slowapi==0.1.8